requirements.txt
haystack-ai
python-dotenv
openai
langchain-community
wikipedia
sentence-transformers>=2.2.0